back_end/requirements.txt
google-generativeai
python-dotenv
pymongo
langchain
faiss-cpu
sentence-transformers
rank-bm25
langchain-community
langchain-huggingface
fastapi
uvicorn
ollama
python-multipart
selenium
beautifulsoup4
webdriver-manager
ultralytics